Tools from toolbox are added to the forms.
There are two forms so far: main form and help form.
Main form will display first, help form will display after clicking the help button in the header of the main form.
All the necessary tools are already present, but they might be altered in the future.
Functions for clicking the buttons are declared, but not defined.
It's late here and I am tired. Let me sleep.